Improve map rendering (a little)

+++ src/Model/Tnt/MapAttribute.cs
@@ -2,7 +2,7 @@
 
 public class MapAttribute
 {
-    public Int16 Elevation { get; set; }
+    public byte Elevation { get; set; }
     public UInt16 TileAnimationIndex { get; set; }
-    public Int16 Unknown { get; set; }
+    public byte Unknown { get; set; }
 }
